move utils to new folder
