+++ main.py
@@ -17,7 +17,7 @@ import torch.backends.cudnn as cudnn
 import torchvision.transforms as transforms
 import torch.nn.functional as F
 
-import myUtils as utils
+import utils.myUtils as utils
 import data_transforms
 from loss_functions import realEPE
 from Test_KITTI import main as test_main
